✨ Added new scripts and a selection manager to the scene
Now we have new scripts that add actual implementation to the selection scripts used in class.
We use these script to create the ray at the cursors position, when we hover over an object, it returns the objects name and layer.
We will extend this to show on ui not only in the inspector.
This will help the players identify planets and check if they visited them yet since some planets look slightly similar to each other.