✨ Implemented PhotoInventoryManager script and reworked code
Set up a temporary image and text on the canvas.
Created a display method to show the photo and the planet's name. Learned how to set a sprite from a texture using Sprite.Create here:https://docs.unity3d.com/6000.0/Documentation/ScriptReference/Sprite.Create.html

Updated the TakeScreenshot method to include a coroutine that waits for the image to be saved, loaded, and displayed in the UI. For loading the image, I referred to Texture2D.LoadImage here:
https://docs.unity3d.com/530/Documentation/ScriptReference/Texture2D.LoadImage.html?utm_source=chatgpt.com