✨ Made A Planet Tracking System
The planets which are further out are very difficult to find normally without a map. I implemented a Planet tracker that will track a planet with a specific tag and point an arrow in the UI to the direction of that planet.
I convert the direction vector to the x-z axis as the game is layed out on that axis not the x-y.